requirements.txt
certifi==2017.4.17
chardet==3.0.3
construct==2.5.3
Django==1.10.7
django-haystack==2.6.1
django-simple-search==1.0.2
future==0.16.0
idna==2.5
pefile==2016.3.28
pysolr==3.3.2
python-ptrace==0.9.2
pytz==2017.2
requests==2.17.3
six==1.10.0
urllib3==1.21.1
